feat: 删除无用依赖
