+++ src/main/java/com/achobeta/domain/resource/model/entity/DigitalResource.java
@@ -2,7 +2,6 @@ package com.achobeta.domain.resource.model.entity;
 
 import com.achobeta.common.base.BaseIncrIDEntity;
 import com.achobeta.common.enums.ResourceAccessLevel;
-import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
@@ -26,6 +25,5 @@ public class DigitalResource extends BaseIncrIDEntity implements Serializable {
 
     private String fileName;
 
-    @TableField(exist = false)
     private static final long serialVersionUID = 1L;
 }
