+++ src/main/java/com/achobeta/domain/resource/service/ResourceService.java
@@ -26,6 +26,8 @@ public interface ResourceService {
 
     void preview(Long code, HttpServletResponse response);
 
+    String gerObjectUrl(Long code);
+
     Long upload(Long userId, MultipartFile file);
 
     List<Long> uploadList(Long userId, List<MultipartFile> fileList);
