+++ src/main/java/com/achobeta/domain/resource/util/ResourceUtil.java
@@ -2,13 +2,9 @@ package com.achobeta.domain.resource.util;
 
 import com.achobeta.common.enums.GlobalServiceStatusCode;
 import com.achobeta.exception.GlobalServiceException;
-import com.achobeta.util.MediaUtil;
-import org.apache.tika.Tika;
 import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.UUID;
 
 /**
@@ -19,8 +15,6 @@ import java.util.UUID;
  * Time: 12:31
  */
 public class ResourceUtil {
-
-    private final static Tika TIKA = new Tika();
 
     public static void checkOriginalName(String originalName) {
         // 判断是否有非空字符以及是否有后缀
@@ -38,22 +32,6 @@ public class ResourceUtil {
     public static String getFileNameSuffix(String originalName) {
         checkOriginalName(originalName);
         return originalName.substring(originalName.lastIndexOf("."));
-    }
-
-    public static String getContentType(InputStream inputStream, String suffix) {
-        try {
-            return TIKA.detect(inputStream, suffix);
-        } catch (IOException e) {
-            throw new GlobalServiceException(e.getMessage(), GlobalServiceStatusCode.RESOURCE_NOT_VALID);
-        }
-    }
-
-    public static String getContentType(byte[] data, String suffix) {
-        try(InputStream inputStream = MediaUtil.getInputStream(data)) {
-            return TIKA.detect(inputStream, suffix);
-        } catch (IOException e) {
-            throw new GlobalServiceException(e.getMessage(), GlobalServiceStatusCode.RESOURCE_NOT_VALID);
-        }
     }
 
     public static String getUniqueFileName(String suffix) {
